requirements.txt
openai
gradio
poe-api
